Fix for single-numbered versioned libraries on OpenBSD
The samba project relies on libraries having only a major number
Author: Volker Lendecke <vl at samba.org>

Note: to contact upstream, use our issue tracker on
https://code.google.com/p/waf/issues/list
Issues that change the behaviour of the system too much or major
feature requests are best implemented in the project itself, or through
the extension mechanism. For example wafadmin/3rdparty/build_file_tracker.py
can be used to trigger rebuilds without changing the fundamental structure
of the build. It is also best to use the latest version of Waf which is 1.8 now.